Replace env var with nix var
Git configs do not support environment variables replacement. But since
those configs are being generated by Nix, it is possible to exploit Nix
as a language, and use the variables during the generation process
itself.
Once the Git configs are written, those are not dynamical any longer, so
subsequent modifications to $XDG_CONFIG_HOME will be ignored.